requirements.txt
flask
psycopg2
flask-Migrate
flask-SQLAlchemy
flask_wtf
wtforms
flask-bootstrap
Flask-Migrate
Pillow
